pyproject.toml
[tool.poetry]
name = "horaro-discord"
version = "0.1.0"
description = ""
authors = ["drago-suzuki58 <dorasuzu58@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.4.0"
pandas = "^2.2.3"
loguru = "^0.7.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
pandas-stubs = "^2.2.3.241009"
faker = "^33.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
